+++ release-notes.rst
@@ -2,10 +2,16 @@ Release notes for chevah.empirical
 ==================================
 
 
+0.13.4 - 13/01/2014
+-------------------
+
+* Fix getSegmentsFromRealPath on Windows.
+
+
 0.13.3 - 13/01/2014
 -------------------
 
-* Fix ILocalFilesyste.openFile declaration.
+* Fix ILocalFilesystem.openFile declaration.
 
 
 0.13.2 - 13/01/2014